Move cmake files into c++ subdir.
The idea behind the directory organization was that we might have official implementations for other languages in other top-level directories, as siblings to the c++ directory. It seems implausible that we'd use CMake as a meta-build-system over all of these.

Another reason for this change is that the release tarball actually contains only the c++ subdirectory. We probably want to include cmake files in the release.