Avoid incorrect library version with BSD sed
BSD sed does not support GNU sed's -r switch, but uses -E instead for
the same functionality. Because -r was used in Makefile.am to set the
SO_VERSION and failed there on platforms with BSD sed (e.g. OS X and
possibly some BSDs, too) SO_VERSION was empty, causing libtool to use
the next given parameter as version number, which was -no-undefined.

Fix this by testing for the regex flag in configure.ac (note that some
implementations of sed support -E but do not document it).

Signed-off-by: Clemens Lang <neverpanic@gmail.com>