Fix UnixEventPort value on 32-bit big endian archs
toRegularSiginfo attempts to convert signalfd_siginfo.ssi_ptr (a 64-bit
integer across all architectures) to a pointer. On 32-bit big endian
architectures, sival_int/sival_ptr will be stored in the high 32-bits
of the ssi_ptr value.

Prior to this change, the value was "lost" as we cast away the high
bits of ssi_ptr in the conversion to a 32-bit pointer.

This fixes AsyncUnixTest::SignalWith{,Pointer}Value on affected archs.