Fix setStructPointer+canonical word mult assume
Previously I assumed that data sections in the input StructReader must
have data sections which were multiples of words. However, struct
upgrade rules allow a StructReader to exist which does not follow that
rule. This commit removes that assumption by working in bytes normally,
and special casing the 1-bit-struct case.