Implement DynamicValue::Reader::as<DynamicValue>()
This is required to allow the compiler to successfully instantiate a
JsonCodec::Handler<DynamicValue, Style::POINTER> using Handler's default
partial specialization for Style::POINTER. Such an instantiation is
currently only used in dead code that clang and gcc eliminate but MSVC
does not, which is why the problem was not caught until now.

Also implemented DynamicValue::Builder::as<DynamicValue>() to preserve
symmetry.