Fix HeapArrayDisposer static initialization order
The Disposer::instance pattern risks the so-called static initialization
order fiasco: that is, the linker gets to define when the
Disposer::instance variable is initialized with respect to when objects in
other translation units are initialized. If another object's dynamic
initializer depends on Disposer::instance, there is a risk of reading
uninitialized data.

A concrete example which motivates this commit is in http-test.c++:

HUGE_STRING is dynamically initialized by strArray(), which returns a
String that MSVC does not NRVO. Thus, String's destructor is called while
still inside a dynamic initializer, which leads to dereferencing
HeapArrayDisposer's vtable pointer (__vfptr), which is null because it
hasn't been initialized yet. This causes a segfault on startup.

Ideally, HeapArrayDisposer::instance would just be a constexpr value.
However, MSVC as of VS2015 does not initialize constexpr __vfptrs,
so we're stuck with the old trick of hiding a static instance inside a
function.

This commit only resolves the problem for http-test.c++ -- there are other
uses of this pattern which carry the same risk, but I wanted to run this by
@kentonv before suggesting it as a general solution.