Fix TwoPartyNetwork/Pipelining test on MSVC.
I can only reproduce this failure on AppVeyor, but the test prints:

capnp\rpc.c++:2886: exception = kj\async-io-win32.c++:399: failed: WSASend(): A request to send or receive data was disallowed because the socket had already been shut down in that direction with a previous shutdown call.

I think this is because WSAESHUTDOWN was considered to be a "bug" exception rather than a "disconnected" exception. It is odd, though, that it doesn't happen when I run locally.