Implement `libtool -release` semantics in CMake
Setting the VERSION property on library targets didn't quite express the
Cap'n Proto library naming policy correctly. The autotools build produces
filenames/sonames of this sort:

  libkj-0.7-dev.so
  libcapnp-0.7-dev.so

But the CMake build produced filenames/sonames of this sort:

  libkj.so.0.7-dev
  libcapnp.so.0.7-dev

As I understand things, this makes some ABI-compatibility guarantees that the
autotools build does not make.

This commit teaches the CMake build to produce the correct filename and
soname. It also copies the installed file to a convenience file, like
`libkj.so`, further emulating the autotools behavior.

Unfortunately, the install-time copy is so ugly that it needs to be factored
out into a function. Since the install-time copy depends on the shared library
already having been installed, the install(TARGETS) command for Cap'n Proto
libraries needs to go into the factored-out function as well, to preserve
correct order of install commands.

I could have named this function kj_install_library() perhaps, but a
kj_add_library() seemed to make sense, subsuming all the add_library() calls. I
don't know what the most elegant way to do it is.

In an unrelated issue, I noticed that autotools-built libraries have their
rpaths set to <libdir>, so I emulated the behavior in CMake. Note that
libkj, the root of the capnp/kj dependency graph, did not have its rpath
set, a nuance I did not duplicate.