Refactor how promises are reduced.
We recently introduced (#698) logit to automatically reduce Promise<RemotePromise<T>> to RemotePromise<T>.

Alas, the implementation strategy does not appear to have worked well in combination with viaradic promises. Having reducePromise() be a method of the promise type is problematic because the SFINAE trick that detects the presence of RemotePromise<T>::reducePromise implicitly instantiates Promsie<T> which potentially recursively loops back into itself, causing some very confusing compiler errors.

I've therefore changed the hook to be a free-standing function based on Koenig lookup rather than a static method. This solves the problem, and in the process I've cleaned some things up as well.