Allow WebSocket to keep sending after close().
In Cloudflare Workers, we've observed applications in the wild that continue to send messages after a Close message. This is incorrect, but it happens, and we end up logging a spurious error as we attempt to proxy the message through.

By removing this restriction, we can now proxy these WebSockets despite the incorrect usage.