Fix bug in byte stream path-shortening.
Apparently, I only initiated capability promise resolution in one of LocalClient's two constructors. Ironically, the other constructor is always the one used by ByteStreams created by ByteStreamFactory, because they are always placed in a CapabilityServerSet. The effect is that ByteStreams couldn't be path-shortened in cases that require promise capabilities to resolve before unwrapping can occur, such as over RPC.

This PR fixes the bug and adds a test covering the RPC use case.