RpcPipeline needs same memoization as QueuedPipeline.
See a few commits ago when QueuedPipeline::getPipelinedCap() was changed to memoize its result. RpcClient needs the same treatment when embargoes are in play, otherwise a second call to getPipelinedCap() for the same capability might miss the embargo set up in the capability returned from a previous call.

This uncovered a test in rpc-test that was actually testing for guarantees that we don't provide...