Suppress bogus valgrind post-exit error.
It seems that running the tests in tls-test.c++ confuse valgrind such that it detects an invalid read after all tests finish. The invalid read looks like a function that is actually legitimately part of the malloc implementation and supposed to be inspecting heap memory outside of allocations. If this is actually a bug, then it seems like a bug in openssl, not our code.