Remove `UnalignedFlatArrayMessageReader`.
This reverts commit 3aa2b2aa02edb1c160b154ad74c08c929a02512a.

This is not safe, even on x86. Unaligned access is undefined behavior in C/C++. Compilers can and do optimize under the assumption that access is aligned. For example, GCC may use SIMD instructions in optimizations, and these do in fact require alignment. In principle, a compiler could also choose to mask away the lowest bits of a pointer that should always be zero when aligned.