Expand comments for Brand.Scope.inherit
The first time I learned about Cap'n Proto generics, and again this weekend while trying to relearn them, I spent a long amount of time trying to understand the meaning of Brand.Scope.inherit. I found the existing doc comment to be unhelpful: what exactly is a "reference point" and what does it mean to inherit from one?

What finally made things click was to think of Brands as type substitutions, and to think of inherit as an identity substitution -- it just replaces the parameters with themselves.

This pull request expands the documentation and adds an example that should make things more clear.