Extend Executor to gracefully handle remote event loop exits.
In practice, we've found that Executor is hard to use correctly if it's possible that the target event loop might exit while other threads are still trying to queue things to its executor. This change makes handling that case more graceful. Now, if an event loop exits without completing a cross-thread task, the requesting thread gets a DISCONNECTED exception (instead of hanging forever or crashing). It's also now possible to hold on to an Executor reference beyond the lifetime of the destination thread, and to query whether its event loop still lives.