Refactor FiberBool to use Disposers.
The freelist is now a Vector of pointers rather than a linked list.

The main reason I'm doing this is to set up for the next change, which will introduce per-CPU freelists. That change will want to be able to do atomic exchanges on raw pointers, so it will be helpful for `FiberPool` to have access to raw pointers rather than `Own`s.