Fix leak when an RPC tail call is canceled.
Consider this situation:
* Client makes call to server.
* The server method tail-calls back to a capability hosted by the client.
* While the server is doing this, the client cancels the original call.

In this situation, the server ends up sending the client messages saying "start this new call, and then use the results to satisfy this other outstanding call you made to me".

However, in the case where the client has already canceled the original call, the code failed to clean up properly. As a result, it ended up never sending a return message for the server -> client call. This would leak some resources, especially including the RPC object to which the tail call was directed on the client side.

In addition to this bug, cancellation was not propagated correctly through these tail calls. This was true whether the cancellation occurred on the client side before or after the server notified it of the tail call.