Fix async exception traces through `promise.catch_()`.
Previously, `catch_()` was implemented in terms of `then()`, passing a specialization of `IdentityFunc` as the non-error handler. But, `TransformPromiseNode` would then use `IdentityFunc` in exception stack traces that passed through the promise chain. This made them not very useful.

Instead, for `catch_()`, we should base the trace address on the error handler. That requires copy/pasting some code but oh well.