Implement magic assert stringification.
When `KJ_ASSERT(foo == bar)` fails, `foo` and `bar`'s actual values will be stringified in the error message. How does it work? We use template magic and operator precedence. The assertion actually evaluates something like this:

    if (auto _kjCondition = kj::_::MAGIC_ASSERT << foo == bar)

`<<` has operator precedence slightly above `==`, so `kj::_::MAGIC_ASSERT << foo` gets evaluated first. This wraps `foo` in a little wrapper that captures the comparison operators and keeps enough information around to be able to stringify the left and right sides of the comparison independently. As always, the stringification only actually occurs if the assert fails.