Minor speedup for miniposix::iovMax
On MacOS, iovMax does an unncessary syscall to determine the maximum
number of iovec entries supported. However, this isn't going to change
during runtime. Move this to a constant that's initialized only once and
remove the `count` argument into iovMax since it's actually superfluous
to the API. The more correct action is to just return
`std::numeric_limits<size_t>::max()` and let the caller figure out what
the value is.